func - (webApiAlt) limpando codigo e alterando regra de negocio

+++ WebApiAlternativa/WebApiAlternativa/Data/Business/Implementations/CategoryBusiness.cs
@@ -1,5 +1,4 @@
-﻿
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using WebApiAlternativa.Data.Repository.Generic;
 using WebApiAlternativa.Entities;
 using WebApiAlternativa.Data.Bussiness;
@@ -9,10 +8,12 @@ namespace WebApiAlternativa.Data.Business.Implementaions
     public class CategoryBusiness : ICategoryBusiness
     {
         private readonly IRepository<Category> _repository;
+        private readonly IRepository<Product> _productRepository;
 
-        public CategoryBusiness (IRepository<Category> repository)
+        public CategoryBusiness (IRepository<Category> repository, IRepository<Product> productRepository)
         {
             _repository = repository;
+            _productRepository = productRepository;
         }
 
         public Category Add (Category category)
@@ -23,9 +24,21 @@ namespace WebApiAlternativa.Data.Business.Implementaions
         {
             return _repository.Update(category);
         }
-        public void Delete(long Id)
+        public bool Delete(long Id)
         {
+            //buscar lista de produtos existentes para procurar relação
+            //Nota: normalmente faria isso no repository de categoria para consultar direto na tabela de produtos,
+            //como a aplicação é pequena faria essa verificação por lista de produtos
+            List<Product> products = _productRepository.GetAll();
+            //procura se há algum produto ligado a categoria
+            Product result = products.Find(product => product.CategoryId.Equals(Id));       
+            if( result != null)
+            {
+                return false;
+            }
+
             _repository.Delete(Id);
+            return true;
         }
         public Category GetById(long Id)
         {
